fix(packaged): move nvm/fnm bin discovery into packaged-runtime sidecars
Keep @open-design/platform focused on generic OS/process primitives by
moving the version-manager PATH discovery helpers (existingDirsUnder,
collectNvmFnmBins) into apps/packaged/src/sidecars.ts where they are
actually consumed by resolvePackagedPathEnv.
