+++ apps/packaged/src/sidecars.ts
@@ -1,4 +1,5 @@
 import { spawn, type ChildProcess } from "node:child_process";
+import { existsSync, readdirSync } from "node:fs";
 import { mkdir, open, type FileHandle } from "node:fs/promises";
 import { createRequire } from "node:module";
 import { homedir } from "node:os";
@@ -22,7 +23,7 @@ import {
   resolveAppIpcPath,
   type SidecarRuntimeContext,
 } from "@open-design/sidecar";
-import { collectNvmFnmBins, createProcessStampArgs, stopProcesses, waitForProcessExit } from "@open-design/platform";
+import { createProcessStampArgs, stopProcesses, waitForProcessExit } from "@open-design/platform";
 
 import type { PackagedNamespacePaths } from "./paths.js";
 
@@ -92,6 +93,29 @@ async function waitForStatus<T>(
 function extractPort(url: string): string {
   const parsed = new URL(url);
   return parsed.port || (parsed.protocol === "https:" ? "443" : "80");
+}
+
+function existingDirsUnder(root: string, segments: string[] = []): string[] {
+  const dirs: string[] = [];
+  try {
+    const entries = readdirSync(root, { withFileTypes: true });
+    for (const entry of entries) {
+      if (!entry.isDirectory()) continue;
+      const full = join(root, entry.name, ...segments);
+      if (existsSync(full)) dirs.push(full);
+    }
+  } catch {
+    // best-effort: directory may not exist or be unreadable
+  }
+  return dirs;
+}
+
+function collectNvmFnmBins(home: string): string[] {
+  return [
+    ...existingDirsUnder(join(home, ".nvm", "versions", "node"), ["bin"]),
+    ...existingDirsUnder(join(home, ".local", "share", "fnm", "node-versions"), ["installation", "bin"]),
+    ...existingDirsUnder(join(home, ".local", "share", "mise", "installs", "node"), ["bin"]),
+  ];
 }
 
 function resolvePackagedPathEnv(basePath = process.env.PATH ?? ""): string {
